screen size - scale

+++ lib/card_testing.dart
@@ -1,3 +1,4 @@
+import 'package:decard/context_extension.dart';
 import 'package:decard/view_source.dart';
 import 'package:flutter/material.dart';
 import 'package:fluttertoast/fluttertoast.dart';
@@ -28,6 +29,7 @@ class _DeCardState extends State<DeCard> {
   static const keyCardStartTime = 'CardStartTime';
 
   final cardWidgetKey = GlobalKey<CardWidgetState>();
+  bool helpAvailability = false;
 
   @override
   Widget build(BuildContext context) {
@@ -46,11 +48,12 @@ class _DeCardState extends State<DeCard> {
                 padding: const EdgeInsets.only(left: 4, right: 4),
                 child: InkWell(
                     onTap: () async {
+                      if (!helpAvailability) return;
                       await ViewContent.navigatorPush(context, widget.child.cardController.card!.pacInfo.path, widget.child.cardController.card!.body.clue, TextConst.txtHelp);
                       final percent = (70 * widget.child.cardController.card!.stat.quality / 100).truncate();
                       _setCostMinusPercent(percent);
                     },
-                    child: const Icon(Icons.live_help, color: Colors.lime)
+                    child: Icon(Icons.live_help, color: helpAvailability? Colors.lime : Colors.grey)
                 ),
               )
             ],
@@ -105,6 +108,8 @@ class _DeCardState extends State<DeCard> {
   Widget _earnedBoxWidget() {
     if (appState.appMode == AppMode.demo) return Container();
 
+    final verticalPadding = inLimit( lineValue( context.scale, 1, 5, 1.9, 0), low: 0, high : 5);
+
     return EventReceiverWidget(
       builder: (_) {
         Color? color;
@@ -115,7 +120,7 @@ class _DeCardState extends State<DeCard> {
         }
 
         final earnedBox = Container(
-          padding: const EdgeInsets.all(5.0),
+          padding: EdgeInsets.only(left: 5, right: 5, top: verticalPadding, bottom: verticalPadding),
           decoration: BoxDecoration(
               color: color,
               border: Border.all(
@@ -182,13 +187,24 @@ class _DeCardState extends State<DeCard> {
   }
 
   Future<void> _selectNextCard() async {
+    helpAvailability = false;
     final ok = await widget.child.cardController.selectNextCard();
     if (ok) {
+      afterSetCard();
       await _saveSelCard();
       setState(() {});
     } else {
       Fluttertoast.showToast(msg: TextConst.txtNoCards);
     }
+  }
+
+  void afterSetCard() {
+    final seconds = 10 + (widget.child.cardController.card!.stat.quality * 50 / 100).round();
+    Future.delayed(Duration(seconds: seconds), (){
+      setState((){
+        helpAvailability = true;
+      });
+    });
   }
 
   Future<void> _setTestCard(int jsonFileID, int cardID, int bodyNum, int startTime) async {
@@ -198,6 +214,7 @@ class _DeCardState extends State<DeCard> {
         await _selectNextCard();
         return;
       }
+      afterSetCard();
       setState(() {});
     } catch (e) {
       await _selectNextCard();
